move style-related methods into base.py
